requirements.txt
openai
instructor
pydantic
pymupdf
pyyaml
